+++ Sandswept/Main.cs
@@ -44,7 +44,7 @@ namespace Sandswept
     {
         public const string ModGuid = "com.TeamSandswept.Sandswept";
         public const string ModName = "Sandswept";
-        public const string ModVer = "0.8.1";
+        public const string ModVer = "0.8.2";
 
         public static AssetBundle MainAssets;
         public static AssetBundle Assets;